Abstract models with `Thing` and `Submission` classes
