+++ api-sdk/src/models/post.ts
@@ -1,52 +1,7 @@
-import Client from "../client";
 import Comment from "./comment";
-import User from "./user";
+import Submission from "./submission";
 
-export default class Post {
-  public id: number;
-  private client: Client;
-  /* Lazily loaded attributes. */
-  public title?: string;
-  public body?: string;
-
-  /* Author */
-
-  public _author?: User;
-
-  get author() {
-    return this._author;
-  }
-
-  set author(newAuthor: any) {
-    let na = newAuthor;
-    if (!isNaN(newAuthor)) {
-      na = new User({ id: newAuthor, client: this.client });
-    } else {
-      na = new User({ client: this.client, ...newAuthor });
-    }
-    this._author = na;
-  }
-
-  /* Comments */
-
-  private _comments?: Comment[];
-
-  get comments() {
-    return this._comments || [];
-  }
-
-  set comments(newComments: any[]) {
-    this._comments = newComments.map((c) =>
-      c instanceof Comment ? c : new Comment({ client: this.client, ...c })
-    );
-  }
-
-  constructor({ id, client, ...data }: any) {
-    this.id = id!!;
-    this.client = client;
-    Object.assign(this, data);
-  }
-
+export default class Post extends Submission {
   public async comment({ body }: { body: string }) {
     const data = await this.client.post("/comments", {
       body,
